Crate pallet_crowdloan_rewards

Crowdloan Rewards Pallet

This pallet is DEPRECATED, the remaining code handle only unclaimed rewards.

Payout Mechanism

The current payout mechanism requires contributors to claim their payouts. Because they are paying the transaction fees for this themselves, they can do it as often as every block, or wait and claim the entire thing once it is fully vested. We could consider auto payouts if we want.

Re-exports

pub use crate::weights::WeightInfo;

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_crowdloan_rewards