pallet_crowdloan_rewards::pallet
Module dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

associate_native_identity

Associate a native rewards_destination identity with a crowdloan contribution.

change_association_with_relay_keys

Change reward account by submitting proofs from relay accounts

claim

Collect whatever portion of your reward are currently vested.

update_reward_address

Update reward address, proving that the caller owns the current native key