pallet_crowdloan_rewards::pallet
Module storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

AccountsPayable

Warning: Doc-Only

ClaimedRelayChainIds

Warning: Doc-Only

EndVestingBlock

Vesting block height at the initialization of the pallet

InitVestingBlock

Vesting block height at the initialization of the pallet

Initialized

Warning: Doc-Only

InitializedRewardAmount

Total initialized amount so far. We store this to make pallet funds == contributors reward check easier and more efficient

TotalContributors

Total number of contributors to aid hinting benchmarking

UnassociatedContributions

Warning: Doc-Only