pallet_crowdloan_rewards::pallet
Trait Config

pub trait Config: Config {

    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type Initialized: Get<bool>;
    type InitializationPayment: Get<Perbill>;
    type MaxInitContributors: Get<u32>;
    type MinimumReward: Get<BalanceOf<Self>>;
    type RewardAddressRelayVoteThreshold: Get<Perbill>;
    type RewardCurrency: Currency<Self::AccountId>;
    type RelayChainAccountId: Parameter + Into<AccountId32> + From<AccountId32> + Ord + Serialize + for<'a> Deserialize<'a>;
    type RewardAddressChangeOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type SignatureNetworkIdentifier: Get<&'static [u8]>;
    type RewardAddressAssociateOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type VestingBlockNumber: AtLeast32BitUnsigned + Parameter + Default + Into<BalanceOf<Self>> + Serialize + for<'a> Deserialize<'a>;
    type VestingBlockProvider: BlockNumberProvider<BlockNumber = Self::VestingBlockNumber>;
    type WeightInfo: WeightInfo;

}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait. Configuration trait of this pallet.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

The overarching event type

type Initialized: Get<bool>

Checker for the reward vec, is it initalized already?

type InitializationPayment: Get<Perbill>

Percentage to be payed at initialization

type MaxInitContributors: Get<u32>

type MinimumReward: Get<BalanceOf<Self>>

The minimum contribution to which rewards will be paid.

type RewardAddressRelayVoteThreshold: Get<Perbill>

A fraction representing the percentage of proofs that need to be presented to change a reward address through the relay keys

type RewardCurrency: Currency<Self::AccountId>

The currency in which the rewards will be paid (probably the parachain native currency)

type RelayChainAccountId: Parameter + Into<AccountId32> + From<AccountId32> + Ord + Serialize + for<'a> Deserialize<'a>

The AccountId type contributors used on the relay chain.

type RewardAddressChangeOrigin: EnsureOrigin<Self::RuntimeOrigin>

type SignatureNetworkIdentifier: Get<&'static [u8]>

Network Identifier to be appended into the signatures for reward address change/association Prevents replay attacks from one network to the other

type RewardAddressAssociateOrigin: EnsureOrigin<Self::RuntimeOrigin>

type VestingBlockNumber: AtLeast32BitUnsigned + Parameter + Default + Into<BalanceOf<Self>> + Serialize + for<'a> Deserialize<'a>

The type that will be used to track vesting progress

type VestingBlockProvider: BlockNumberProvider<BlockNumber = Self::VestingBlockNumber>

The notion of time that will be used for vesting. Probably either the relay chain or sovereign chain block number.

type WeightInfo: WeightInfo

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.